# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Safety-first multilingual rural health assistant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type RiskLevel = typeof RiskLevel[keyof typeof RiskLevel];


export const RiskLevel = {
  low: 'low',
  medium: 'medium',
  emergency: 'emergency',
} as const;

export type Language = typeof Language[keyof typeof Language];


export const Language = {
  en: 'en',
  hi: 'hi',
  mr: 'mr',
} as const;

export type ProfileAccessibility = { [key: string]: unknown };

export interface Profile {
  preferredLanguage: Language;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  sex?: string | null;
  /** @nullable */
  pregnancyStatus?: string | null;
  /** @nullable */
  emergencyContactName?: string | null;
  /** @nullable */
  emergencyContactPhone?: string | null;
  allergies?: string[];
  chronicConditions?: string[];
  currentMedications?: string[];
  accessibility?: ProfileAccessibility;
}

export type ProfileInputAccessibility = { [key: string]: unknown };

export interface ProfileInput {
  preferredLanguage?: Language;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  sex?: string | null;
  /** @nullable */
  pregnancyStatus?: string | null;
  /** @nullable */
  emergencyContactName?: string | null;
  /** @nullable */
  emergencyContactPhone?: string | null;
  allergies?: string[];
  chronicConditions?: string[];
  currentMedications?: string[];
  accessibility?: ProfileInputAccessibility;
}

export type AssessmentInputSymptomData = { [key: string]: unknown };

export interface AssessmentInput {
  language: Language;
  /**
     * @minLength 1
     * @maxLength 2000
     */
  chiefComplaint: string;
  symptomData: AssessmentInputSymptomData;
  riskLevel?: RiskLevel;
  riskReasons?: string[];
  redFlags?: string[];
  recommendedAction?: string;
  aiSummary?: string;
}

export type Assessment = AssessmentInput & {
  id: string;
  createdAt: string;
};

export type SymptomAnalysisInputContext = { [key: string]: unknown };

export interface SymptomAnalysisInput {
  /**
     * @minLength 1
     * @maxLength 4000
     */
  message: string;
  language: Language;
  context?: SymptomAnalysisInputContext;
}

export interface SymptomAnalysis {
  riskLevel: RiskLevel;
  redFlags: string[];
  reasoningSummary: string;
  recommendedAction: string;
  seekEmergencyCare: boolean;
  followUpQuestions: string[];
  extractedSymptoms?: string[];
}

export interface SummaryInput {
  language: Language;
  assessment: AssessmentInput;
  profile?: Profile;
}

export interface DoctorSummary {
  disclaimer: string;
  text: string;
  riskLevel: RiskLevel;
  unknownInformation?: string[];
}

export type ReminderInputType = typeof ReminderInputType[keyof typeof ReminderInputType];


export const ReminderInputType = {
  medication: 'medication',
  appointment: 'appointment',
  'follow-up': 'follow-up',
} as const;

export type ReminderInputSchedule = { [key: string]: unknown };

export interface ReminderInput {
  /**
     * @minLength 1
     * @maxLength 120
     */
  title: string;
  type: ReminderInputType;
  /** @nullable */
  dosage?: string | null;
  schedule: ReminderInputSchedule;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  endDate?: string | null;
  /** @nullable */
  notes?: string | null;
}

export type Reminder = ReminderInput & {
  id: string;
};

export interface Facility {
  id: string;
  name: string;
  type: string;
  address: string;
  /** @nullable */
  distanceKm?: number | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  openingInfo?: string | null;
  /** @nullable */
  emergencyAvailable?: boolean | null;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  source?: string;
}

export type DashboardStats = {
  assessments: number;
  activeReminders: number;
};

export interface Dashboard {
  recentAssessments: Assessment[];
  reminders: Reminder[];
  stats: DashboardStats;
}

export type GetNearbyFacilitiesParams = {
latitude?: number;
longitude?: number;
type?: string;
emergency?: boolean;
};