# lib/api-zod/src/generated/types/assessmentInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Safety-first multilingual rural health assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { AssessmentInputSymptomData } from './assessmentInputSymptomData';
import type { Language } from './language';
import type { RiskLevel } from './riskLevel';

export interface AssessmentInput {
  language: Language;
  /**
     * @minLength 1
     * @maxLength 2000
     */
  chiefComplaint: string;
  symptomData: AssessmentInputSymptomData;
  riskLevel?: RiskLevel;
  riskReasons?: string[];
  redFlags?: string[];
  recommendedAction?: string;
  aiSummary?: string;
}