# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Safety-first multilingual rural health assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { Assessment } from './assessment';
import type { DashboardStats } from './dashboardStats';
import type { Reminder } from './reminder';

export interface Dashboard {
  recentAssessments: Assessment[];
  reminders: Reminder[];
  stats: DashboardStats;
}