# lib/api-zod/src/generated/types/doctorSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Safety-first multilingual rural health assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { RiskLevel } from './riskLevel';

export interface DoctorSummary {
  disclaimer: string;
  text: string;
  riskLevel: RiskLevel;
  unknownInformation?: string[];
}