# lib/api-zod/src/generated/types/facility.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Safety-first multilingual rural health assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface Facility {
  id: string;
  name: string;
  type: string;
  address: string;
  /** @nullable */
  distanceKm?: number | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  openingInfo?: string | null;
  /** @nullable */
  emergencyAvailable?: boolean | null;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  source?: string;
}