# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Safety-first multilingual rural health assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { Language } from './language';
import type { ProfileAccessibility } from './profileAccessibility';

export interface Profile {
  preferredLanguage: Language;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  sex?: string | null;
  /** @nullable */
  pregnancyStatus?: string | null;
  /** @nullable */
  emergencyContactName?: string | null;
  /** @nullable */
  emergencyContactPhone?: string | null;
  allergies?: string[];
  chronicConditions?: string[];
  currentMedications?: string[];
  accessibility?: ProfileAccessibility;
}