# lib/api-zod/src/generated/types/reminder.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Safety-first multilingual rural health assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { ReminderInput } from './reminderInput';

export type Reminder = ReminderInput & {
  id: string;
};