# lib/api-zod/src/generated/types/riskLevel.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Safety-first multilingual rural health assistant API
 * OpenAPI spec version: 0.1.0
 */

export type RiskLevel = typeof RiskLevel[keyof typeof RiskLevel];


export const RiskLevel = {
  low: 'low',
  medium: 'medium',
  emergency: 'emergency',
} as const;